+++ src/Tools/SourceLocalization/__init__.py
@@ -1,6 +1,7 @@
 """Source localization tools using (s/e)LORETA."""
 
 from .eloreta_gui import SourceLocalizationWindow
-from .eloreta_runner import run_source_localization, view_source_estimate
+from .runner import run_source_localization
+from .visualization import view_source_estimate
 
 __all__ = ["SourceLocalizationWindow", "run_source_localization", "view_source_estimate"]
